requirements-dev.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.12.0
idna==2.10
mypy-extensions==0.4.3
pathspec==0.8.0
python-decouple==3.3
pytz==2020.1
regex==2020.9.27
requests==2.24.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
